# src/docs/swagger.yaml
definitions:
...
  api.CreateGameRequest:
    properties:
      opponent_type:
        type: string
    type: object
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  api.GameRequest:
    properties:
      board:
        items:
          items:
            type: integer
          type: array
        type: array
    type: object
  api.GameResponse:
    properties:
      board:
        items:
          items:
            type: integer
          type: array
        type: array
      created_at:
        type: string
      game_status:
        type: string
      id:
        type: string
      player_o_id:
        type: string
      player_x_id:
        type: string
    type: object
...
  api.UserResponse:
    properties:
      login:
        type: string
      user_id:
        type: string
    type: object
...
  domain.LiederBoard:
    properties:
      id:
        type: string
      win_rate:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /game:
    post:
      consumes:
      - application/json
      description: Инициализирует новую игру 3х3 с выбором оппонента (ИИ или человек).
        Требует авторизацию.
      parameters:
      - description: Принимает строку ai или human
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.CreateGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.GameResponse'
        "400":
          description: Неверный формат тела запроса или невалидный opponent_type
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при сохранении в БД
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создание новой игровой комнаты
      tags:
      - game
  /game/{current_game_UUID}:
    get:
      description: Возвращает текущее состояние игровой сессии (доску, статус, ID
        игроков) по её UUID.
      parameters:
      - description: UUID игровой сессии
        in: path
        name: current_game_UUID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.GameResponse'
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Игровая сессия не найдена
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о конкретной игре
      tags:
      - game
    post:
      consumes:
      - application/json
      description: |-
        Принимает новое состояние игрового поля от авторизованного пользователя.
        Проверяет очередность хода, валидирует корректность изменения клеток.
        Если оппонент — ИИ, то сервер автоматически рассчитывает и совершает ответный ход.
      parameters:
      - description: UUID игровой сессии
        in: path
        name: current_game_UUID
        required: true
        type: string
      - description: Новое состояние игрового поля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.GameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ход успешно принят (и выполнен ответный ход ИИ, если применимо)
          schema:
            $ref: '#/definitions/api.GameResponse'
        "400":
          description: Неверный формат UUID, сломанный JSON, игра окончена или еще
            не началась
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "403":
          description: Ход не в свою очередь или пользователь не является участником
            этой игры
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Игровая сессия не найдена в базе данных
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "422":
          description: Ошибка валидации доски (изменено более 1 клетки, затерты чужие
            фигуры)
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при расчетах ИИ или сохранении в
            БД
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Совершение хода в игре
      tags:
      - game
  /game/{current_game_UUID}/join:
    post:
      description: Позволяет авторизованному пользователю занять свободное место (Нолика)
        в комнате со статусом ожидания.
      parameters:
      - description: UUID игровой сессии
        in: path
        name: current_game_UUID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешное подключение, статус игры изменен на PlayerXTurn
          schema:
            $ref: '#/definitions/api.GameResponse'
        "400":
          description: Неверный формат UUID или бизнес-ошибка (комната заполнена,
            попытка войти в свою комнату)
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Присоединение к существующей игре
      tags:
      - game
  /games/available:
    get:
      description: Возвращает список всех созданных игровых комнат, находящихся в
        статусе ожидания второго игрока (WaitingForPlayers).
      produces:
      - application/json
      responses:
        "200":
          description: Список доступных комнат (может быть пустым [])
          schema:
            items:
              $ref: '#/definitions/api.GameResponse'
            type: array
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при запросе к БД
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение списка доступных игр
      tags:
      - game
  /games/history:
    get:
      description: Возвращает список всех завершенных игр (победа, поражение, ничья),
        в которых участвовал текущий пользователь. Требует авторизацию.
      produces:
      - application/json
      responses:
        "200":
          description: Список завершенных игр
          schema:
            items:
              $ref: '#/definitions/api.GameResponse'
            type: array
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера при чтении из БД
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение истории завершенных игр пользователя
      tags:
      - game
  /games/leaderboard:
    get:
      description: Возвращает топ N игроков с наилучшим соотношением побед к поражениям
        и ничьим.
      parameters:
      - description: Количество игроков в выборке (по умолчанию 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.LiederBoard'
            type: array
        "400":
          description: Некорректное значение параметра limit
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение таблицы лидеров
      tags:
      - game
...
  /users/{user_uuid}:
    get:
      description: Возвращает публичный профиль пользователя (ID и логин) по его UUID.
      parameters:
      - description: UUID пользователя
        in: path
        name: user_uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Неверный формат UUID
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Пользователь не авторизован
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получение информации о пользователе
      tags:
      - user
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {accessToken}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"